requirements.txt
multidict
discord.py
requests